cf: Disable pam if --disable-shared
If we are configured with --enable-kauth and --disable-shared, 'make
install/dest' currently fails for pam, since we never generate the
shared library files:

    make[3]: Entering directory '.../src/pam'
    if [ "xyes" = "xyes" ]; then \
            /usr/bin/install -c -d .../amd64_linux26/dest/lib; \
            /usr/bin/install -c -m 644 .libs/pam_afs.so .../amd64_linux26/dest/lib; \
            /usr/bin/install -c -m 644 .libs/pam_afs.krb.so .../amd64_linux26/dest/lib; \
    fi
    /usr/bin/install: cannot stat '.libs/pam_afs.so': No such file or directory
    /usr/bin/install: cannot stat '.libs/pam_afs.krb.so': No such file or directory
    make[3]: *** [dest] Error 1

Building pam plugins doesn't make sense if shared libraries are
disabled, so just turn off pam if --disable-shared is specified, similar
to what we do if pam headers aren't available.

While we're here, slightly restructure the code for setting HAVE_PAM, to
make it easier to add pam-disabling conditions like this.

Change-Id: I7d1f98267ba342dbd7290de40a07da904e65dbe3
Reviewed-on: https://gerrit.openafs.org/15998
Reviewed-by: Ben Huntsman <ben@huntsmans.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>