AIX: Fix rx_knet.c return types
When compiling the kernel extension on AIX 7.2+ using the CLANG-based
Open XL C 17.1+ compiler, the build of src/rx/AIX/rx_knet.c throws two
errors:

    /home/build/openafs/src/rx/AIX/rx_knet.c:167:6: error: non-void function 'rxk_RX_input' should return a value [-Wreturn-type]
                return;
                ^
    ...
    /home/build/openafs/src/rx/AIX/rx_knet.c:239:6: error: non-void function 'rxk_isr' should return a value [-Wreturn-type]
                return;
                ^

To fix these, change rxk_RX_input() to be 'void', and make it static
while we're here. Change rxk_isr() to return 0 (and declare it properly
to return int); it's not clear what the return value should be, but this
function is passed to add_input_type(), which expects a function that
returns int.

Change-Id: I436697e749d91f7586e1c299d4b39a84c2031684
Reviewed-on: https://gerrit.openafs.org/16568
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Ben Huntsman <ben@huntsmans.net>
Tested-by: Ben Huntsman <ben@huntsmans.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>