lwp: Link selclient and selserver with XLIBS
Commit e4d8329619f3 (lwp: Build src/lwp/test) fails to build on Solaris
11.3 or older due to undefined socket and resolver API symbols,
including recv, send, gethostbyname, socket, etc:

  Undefined                       first referenced
   symbol                             in file
  recv                                selsubs.o
  send                                selsubs.o
  gethostbyname                       selclient.o
  socket                              selclient.o
  connect                             selclient.o
  ld: fatal: symbol referencing errors

This is because on Solaris, these have historically been supplied via
-lsocket, -lresolv, and other libraries as defined in the OpenAFS
variable XLIBS.

Add XLIBS to the recipes for selclient and selserver to avoid these
errors when building on Solaris 11.3 or older.

On Solaris 11.4, XLIBS is not required to build because the APIs
formerly found via XLIBS are now largely provided in libc.so instead,
with libsocket.so and libresolv.so maintained for compatibility as
filters on libc.so.

Change-Id: I19a5cca20bd9ac335cd0e6424ed8aeb85d574bc8
Reviewed-on: https://gerrit.openafs.org/16033
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>