AIX: Avoid -H8 when linking libafs
Since OpenAFS 1.0, we've passed -H8 to the linker when building our
kernel extensions in src/libafs. This causes various sections in
generated binary to be 8-byte aligned, but it's not clear why it was
ever passed.

When building the kernel extension on AIX using the new CLANG-based Open
XL C 17.1+ compilers, this causes the following warning:

    ld: 0711-356 WARNING: The .text section may not be aligned properly.

The kernel extension still builds, but fails to load into the kernel.

Remove the -H8 option to use the default behavior, which causes the
sections in the resulting binary to be word-aligned. This gets rid of
the warning message, and allows the kernel extension to load.

Change-Id: I8aafaed3a620f19953b7730aae7c2b901979d1f3
Reviewed-on: https://gerrit.openafs.org/16572
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Ben Huntsman <ben@huntsmans.net>
Tested-by: Ben Huntsman <ben@huntsmans.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>