AIX: Return real kcred_setpag() error from setpag
Currently, in AIX's setpag(), if kcred_setpag() fails, our 'code' will
be -1 (since kcred_setpag() returns -1 on error; the real error is in
getuerror()). And so, we'll return -1 from setpag(), whereas for all
other errors, we return the error code directly like we do for most
openafs functions.

In afs_osi_pag.c:AddPag(), the most common caller of setpag(), we'll
then call setuerror(-1), which is clearly incorrect, overwriting the
actual error code set by kcred_setpag().

To fix this, if kcred_setpag() returns an error, get the real error code
with getuerror(). Move the final setuerror() call in setpag() to be
called for all platforms, instead of hidden inside an #ifdef; this way,
the error-handling code at the end of setpag() is the same for all
platforms, and a future commit can consolidate the error-handling logic
inside a 'goto done'-style destructor.

If kcred_setpag() returns an error, the new setuerror() call will be
redundant, but that's harmless.

Thanks for mvitale@sinenomine.net for pointing out some relevant
problems here.

Change-Id: I6eee00961d0953e8299d7d5e678c53989bab7487
Reviewed-on: https://gerrit.openafs.org/16087
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>