budb: Desupport buserver -resetdb option
The buserver -resetdb option has not functioned since OpenAFS 1.0 and
the code present in the tree for that option has serveral bugs and
triggers static analyzer warnings.

Instead of trying to fix those bugs and implement this misfeature, just
log a warning message to the buserver log and document this option is
ignored.

Change-Id: Ia2d6fa47cdb7add6f8044e9d7b88bc856ae46b6b
Reviewed-on: https://gerrit.openafs.org/16108
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>