tests: Swap incorrect got and expected args in Perl tests
Swap the "got" and "expected" arguments where the order was incorrect in
the Perl TAP tests.  This does not effect the test but makes the output
make more sense when a test fails, since the Perl Test::More is()
function prints the "got" and "expected" labels when a test fails.

For example:

    not ok 3 - multi threaded client ran succesfully
    #   Failed test 'multi threaded client ran succesfully'
    #   at ./perf-t line 58.
    #          got: '0'
    #     expected: '254'

The c-tap is_*() functions print failed results with "left" and "right"
labels instead of "got" and "expected", so we only need to get the order
correct in the tests implemented in Perl, so this commit changes just
the Perl tests.

Change-Id: Ieeff3b7f4803f70ec5452e0bdce3e4b22c0df073
Reviewed-on: https://gerrit.openafs.org/16119
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>