DARWIN: Set parent of volume root vnodes
Vnodes that lack a parent can lead to inaccurate ENOENT errors on macOS.
To prevent this issue from occurring, ensure that the parent of volume
root vnodes is always set when evaluating them.

Thanks to gangovind@in.ibm.com, who originally brought attention to some
of the side effects of this problem.

This commit is the product of collaborative efforts by
mbarbosa@sinenomine.net, adeason@sinenomine.net, and kaduk@mit.edu.

Change-Id: I97024989de294a650497237d95afd3123b80d4e6
Reviewed-on: https://gerrit.openafs.org/16032
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>