afs: Use constants for vrequest.volumeError
A few callers were setting struct vrequest.volumeError to '1', even
though the values of volumeError are named constants (e.g., 1 is
VOLMISSING). Make sure all callers use the named constants.

In afs_CopyError(), we were also overwriting the copied volumeError
value with 1; instead, use the actual error from the source vrequest.
The only caller of afs_CopyError() is afs_NewVolumeByName(), so the only
practical effect of this should be minor, and just potentially affects
the error code returned from an operation that hits an error in
afs_NewVolumeByName().

Change-Id: Ibd6d6be0e2f7e4ebbe9466aee55734ca984b16f5
Reviewed-on: https://gerrit.openafs.org/16143
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>