DARWIN: Avoid libafs .c.o rule
Since commit 9f94892f8d (darwin-x86-and-leopard-20060309),
MakefileProto.DARWIN.in specifies a pattern rule .c.o to build object
files in addition to specifying CRULE_NOOPT and CRULE_OPT. The commands
in the .c.o rule are effectively identical to CRULE_NOOPT, and all of
our libafs objects are built with CRULE_NOOPT or CRULE_OPT, so the
pattern rule should be unnecessary. Get rid of it, to make it easier to
change our CRULE_* logic.

However, to make sure we don't accidentally build something differently
with the pattern rule, make the .c.o rule just fail immediately. This
shouldn't be needed, but do this just in case, since our DARWIN build
rules can be complex and hard to debug.

Change-Id: I126ffd864c7012f1ec7fcb8a20ef021adcfc20be
Reviewed-on: https://gerrit.openafs.org/16204
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>