libafs: Convert $(CFLAGS-foo.o) to $(CFLAGS_foo.o)
Since commit 9f94892f8d (darwin-x86-and-leopard-20060309), our libafs
makefiles have used a convention of CFLAGS-foo.o to specify per-target
CFLAGS. This differs from the convention used in the rest of the tree
and the LINUX-specific libafs makefiles, which use CFLAGS_foo.o to
specify per-target CFLAGS. In addition to being confusing, this has also
resulted in some per-target CFLAGS being duplicated:
MakefileProto.LINUX.in specifies CLFAGS_foo.o, and Makefile.common.in
specifies CFLAGS-foo.o for the same thing.

Change all of our CFLAGS-foo.o references to use CFLAGS_foo.o instead,
to match the rest of the tree.

Remove all of the CFLAGS_foo.o definitions in MakefileProto.LINUX.in,
and just specify the per-target CFLAGS in Makefile.common.in instead.
Most of these were already specified in Makefile.common.in with
identical contents, but there were a few differences:

- Makefile.common.in was specifying CFLAGS-opr-rbtree.o, but the target
  is opr_rbtree.o. Change this to CFLAGS_opr_rbtree.o.

- MakefileProto.LINUX.in was specifying CFLAGS for evp-kernel.o and
  random.o; neither of those exist, so just remove the CFLAGS.

- The CFLAGS for a few files differed in specifying an include path of
  -I$(TOP_INCDIR)/hcrypto vs -I$(TOP_SRCDIR)/external/heimdal/hcrypto.
  The TOP_INCDIR path contains a subset of the headers in the TOP_SRCDIR
  path, but otherwise the headers are the same. Add or modify those
  CFLAGS in Makefile.common.in only when needed: md5.o, sha.o, and
  sha256-kernel.o need the TOP_SRCDIR path for hash.h, and aes.o and
  rijndael-alg-fst.o need the TOP_SRCDIR path for rijndael-alg-fst.h.
  Otherwise leave the existing paths in Makefile.common.in alone; a few
  effectively have changed on LINUX from
  -I$(TOP_SRCDIR)/external/heimdal/hcrypto to -I$(TOP_INCDIR)/hcrypto,
  but they don't need the TOP_SRCDIR path.

- evp.o and evp-algs.o had extra warning inhibitions in
  MakefileProto.LINUX.in; add those to Makefile.common.in.

- Makefile.common.in was passing some extra flags in the build rule for
  afs_pag_user.o; change it to use CFLAGS_afs_pag_user.o to match what
  MakefileProto.LINUX.in had.

- MakefileProto.LINUX.in was passing -DNO_CONFIG_H for
  rijndael-alg-fst.o, but this symbol doesn't seem to be used by
  anything. Remove it.

While we're here, also move our CFLAGS_foo.o definitions to go before
the rule for foo.o, instead of after, if they are specified right next
to the rule for foo.o.

Change-Id: If9b4ceceb0e6a6caf6c578d792185f6aba0e884a
Reviewed-on: https://gerrit.openafs.org/16209
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>