DARWIN: Refactor libafs KOPTS_arch vars
The current arch-related KOPTS_* vars in MakefileProto.DARWIN.in have
become somewhat unreadable. Split out some common flags into
KOPTS_common, and move the remaining flags to be in separate lines, to
make it easier to see what's going on.

Remove the unused KDEFS and DBUG vars while we're here.

Change-Id: I29b2fcd726752432fc08303065334c21e7f0eabd
Reviewed-on: https://gerrit.openafs.org/16212
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>