DARWIN: Convert ARCH_foo to autoconf vars
Currently, src/lwp/Makefile.in and src/libafs/MakefileProto.DARWIN.in
have some DARWIN build logic that needs to know which architectures
we're building for (i386, x86_64, arm64, etc). The logic for these is
rather different but both are hard to read, and both look at ARCHFLAGS
to get the arches to build for, with a default fallback. The default
case is also calculated completely differently ('uname -m' vs 'cc -dM
-E').

The current logic in MakefileProto.DARWIN.in is also a bit odd in that
the ARCH_foo vars can be set to "yes\nyes" (not just "yes"). For example
(following the logic for the current ARCH_amd64 var):

    $ ARCHFLAGS='-arch x86_64 -arch arm64'
    $ echo "${ARCHFLAGS}" | grep -q -w x86_64 && echo yes \
	|| echo ${ARCHFLAGS} | egrep -q -w -v 'ppc|i386|x86_64|arm64' \
	&& uname -m | grep -q -w x86_64 && echo yes
    yes
    yes

This happens because short-circuited logic effectively groups the
commands like this:

    $ ((echo "${ARCHFLAGS}" | grep -q -w x86_64 && echo yes) \
	|| (echo ${ARCHFLAGS} | egrep -q -w -v 'ppc|i386|x86_64|arm64')) \
    && (uname -m | grep -q -w x86_64 && echo yes)
    yes
    yes

And not like this, which is probably what was intended:

    $ (echo "${ARCHFLAGS}" | grep -q -w x86_64 && echo yes) \
	|| (echo ${ARCHFLAGS} | egrep -q -w -v 'ppc|i386|x86_64|arm64' \
	    && uname -m | grep -q -w x86_64 && echo yes)
    yes

This doesn't cause any problems right now, but this shows how
error-prone this pile of shell can be.

Try to unify these areas of the build by changing both to use ARCH_foo
vars that are AC_SUBST'd by autoconf during configure. Have configure
look at ARCHFLAGS, and set ARCH_foo to 'yes' if '-arch foo' is in
ARCHFLAGS. Set the default arch based on $host.

Also rename ARCH_amd64 to ARCH_x86_64, just to make the autoconf logic
easier (since the ARCHFLAGS argument is '-arch x86_64'). Rename ARCH_x86
to ARCH_i386 for the same reason.

Change-Id: I63925f5bc23a7b7197b6328acdb37cf29d2d6e23
Reviewed-on: https://gerrit.openafs.org/16214
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>