DARWIN: Ensure -arch arm64e for MODLD on arm64
Currently, we translate '-arch arm64' into '-arch arm64e' in ARCHFLAGS
when linking our kernel module in MODLD. But if we don't have ARCHFLAGS
set at all, we'll use the default arch, which is 'arm64' on arm64 (not
arm64e). So, if we build without ARCHFLAGS set, the resulting kernel
module won't load, because it's the wrong arch.

To fix this, if ARCHFLAGS is empty, set it to '-arch arm64e' explicitly
for arm64.

Reviewed-on: https://gerrit.openafs.org/16211
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
(cherry picked from commit f5a16cfb7b19efb2858d1fbd82fbaa2573f28371)

Change-Id: I8703b27abd0ed8da5ae51c55ed6f97ba66125a84
Reviewed-on: https://gerrit.openafs.org/16512
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: BuildBot <buildbot@rampaginggeek.com>