DARWIN: Ensure -arch arm64e for MODLD on arm64
Currently, we translate '-arch arm64' into '-arch arm64e' in ARCHFLAGS
when linking our kernel module in MODLD. But if we don't have ARCHFLAGS
set at all, we'll use the default arch, which is 'arm64' on arm64 (not
arm64e). So, if we build without ARCHFLAGS set, the resulting kernel
module won't load, because it's the wrong arch.

To fix this, if ARCHFLAGS is empty, set it to '-arch arm64e' explicitly
for arm64.

Change-Id: If1cb1bc539e2c61a166d89f3d827766751dd5a2a
Reviewed-on: https://gerrit.openafs.org/16211
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>