ptserver: Rename db_verify to prdb_check
Since before OpenAFS 1.0, the prdb_check tool has been built as
db_verify in the build tree and then renamed to prdb_check when
installed with 'make install' (or 'make dest').  This was probably done
long ago to be more consistent with the vldb_check program name.

Rename the main source file for this tool from db_verify.c to
prdb_check.c and build the program as prdb_check to avoid having two
different names for the same program.

Change the "whoami" global to "prdb_check" as well, so the tool will
identify itself as "prdb_check" in error messages, instead of
"db_verify".

Change-Id: Ia534eb239972448cd722104b67b73707c820bd01
Reviewed-on: https://gerrit.openafs.org/16217
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: Michael Meffie <mmeffie@sinenomine.net>