afs: Log addr with "RXAFS_GetCapabilities failed"
If our RXAFS_GetCapabilities() call fails in afs_GetCapabilities() for
any reason besides RXGEN_OPCODE, we'll log a message, but the log
message doesn't say which server we were trying to contact. We usually
log a "Lost contact with file server" message right before this, so
which server it's for is usually intuitive, but that won't always happen
(if we get a positive error code, for example).

To make this message more useful, log the server address with the rest
of the message.

Change-Id: I178d3f8ff234bf89fad38d3fb84c87757d2e1f7c
Reviewed-on: https://gerrit.openafs.org/16245
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>