viced: Remove connSuccess
Currently, a couple of our callers of multi_Rx process the results using
a 'connSuccess' var to record the conn of the first successful call. We
also store success_i, the index for the successful call in our various
arrays. This creates some ambiguity (should we reference
conns[success_i], or connSuccess?), and means we need to make sure we
don't free connSuccess, since it's been stored elsewhere.

To make this a little less error prone, refactor these functions to
refer to the success conn as just conns[success_i], and remove the
connSuccess var. Set conns[success_i] to NULL to more conventionally
indicate that it shouldn't be freed.

Change-Id: I403c1de1e80307f24986210ae445b351aa5fadbc
Reviewed-on: https://gerrit.openafs.org/16274
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>