tests: Check if '$command help' fails
Currently, we have tests to check that we have manpages for various
commands and subcommands; the *-man-t tests, e.g., ptserver/pts-man-t.
These tests run, for example, 'pts help' to get the list of subcommands
to check for. If running 'pts help' fails because of some low-level
issue (such as, some runtime linking error), the manpage tests fail in a
rather confusing way:

    ptserver/pts-man

    1..8
    not ok 1 - existence of man page for pts_'dump
    #   Failed test 'existence of man page for pts_'dump'
    #   in .../tests/tests-lib/perl5/mancheck_utils.pm at line 90.
    not ok 2 - existence of man page for pts_0509-130
    #   Failed test 'existence of man page for pts_0509-130'
    #   in .../tests/tests-lib/perl5/mancheck_utils.pm at line 90.

This is because we don't check if the 'pts help' command failed, and we
assume the output we have captured is the actual help output. But our
command might fail with this output, for example, on AIX 6:

    exec(): 0509-036 Cannot load program .../pts because of the following errors:
      0509-130 Symbol resolution failed for pts because:
      0509-136   Symbol lsetpag (number 115) is not exported from
             dependent module /unix.
      0509-136   Symbol lpioctl (number 116) is not exported from
             dependent module /unix.
      0509-192 Examine .loader section symbols with the
           'dump -Tv' command.

We currently try to interpret that as a list of subcommands, which leads
to nonsensical results.

To make it easier to see what's happening in the face of such errors,
just capture the whole command output in an array, and print out the
output via diag() if the command fails, and bail out before scanning for
subcommand manpages.

Change-Id: I3cf5277f7b98db8c066551cf786831d2fbfa443e
Reviewed-on: https://gerrit.openafs.org/16282
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>