libtool: Remove static lib before building
On DARWIN, if we try to build a static library (for example, libuafs.a),
and the library file already exists, and we are building for multiple
architectures via ARCHFLAGS, linking the library fails:

    /bin/sh ../../libtool --quiet --mode=compile --tag=CC  gcc  -arch x86_64 -arch arm64 [...] -o xdr_len.lo
    /bin/sh ../../libtool --quiet --mode=link --tag=CC  gcc -static  -g -Os -o libuafs.a [...]
    ar: libuafs.a is a fat file (use libtool(1) or lipo(1) and ar(1) on it)
    ar: libuafs.a: Inappropriate file type or format
    make[3]: *** [libuafs.a] Error 1

Normally this doesn't happen during a single build, but can happen if we
rebuild the tree after, for example, editing a source file or other
dependency, and don't 'make clean'.

To allow for such rebuilds, remove the target library file before
creating it again. Do this in LT_LDLIB_static, so we don't need to
update all callers individually.

Change-Id: Ifa6e2693294c17ed03fe4e4c4980bf935da10b51
Reviewed-on: https://gerrit.openafs.org/16293
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>