Convert HAVE_STRUCT_LABEL_SUPPORT to AFS_STRUCT_INIT
The OpenAFS coding style allows the use of designated initializers for
structs, however not all supported platforms have compilers that support
it.

The preprocessor define HAVE_STRUCT_LABEL_SUPPORT has been available for
use so structure initialization can be set up to support compilers that
do have designated initializers support.

The typical use is:

    struct foo x = {
    #ifndef HAVE_STRUCT_LABEL_SUPPORT
        val1,
        val2,
        ...
    #else
        .mem1 = val1,
        .mem2 = val2,
        ...
    #endif
    };

This results in extra lines of code where errors can easily be
introduced.

Create a macro, AFS_STRUCT_INIT, that uses designated initializers when
available, so the above example would be:

    struct foo x = {
        AFS_STRUCT_INIT(.mem1, val1),
        AFS_STRUCT_INIT(.mem2, val2),
        ...
    };

Convert the initialization of structures that are using the
HAVE_STRUCT_LABEL_SUPPORT define to use AFS_STRUCT_INIT.

Note, there is still a requirement that the order of initializers match
the order of the elements within the structure, but it should be easier
to verify that the initializers are in the proper order.

Use a consistent alignment, and add trailing comma on the last
element.

There are no functional changes made by this commit.

Change-Id: I7c51de0c5decaa8331b8366d26a7e80d64617fb0
Reviewed-on: https://gerrit.openafs.org/16289
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>